requirements.txt
argcomplete @ file:///home/conda/feedstock_root/build_artifacts/argcomplete_1682324182515/work
beautifulsoup4 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_bczxn9gqkq/croot/beautifulsoup4-split_1681493044307/work
biopython @ file:///Users/runner/miniforge3/conda-bld/biopython_1720014913437/work
boltons @ file:///Users/ec2-user/ci_py311/boltons_1678396074811/work
brotlipy==0.7.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
cffi @ file:///Users/ec2-user/ci_py311/cffi_1678315312775/work
chardet @ file:///Users/ec2-user/ci_py311/chardet_1678326102860/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cmake==3.27.0
conda==23.3.1
conda-build==3.24.0
conda-package-handling @ file:///Users/ec2-user/ci_py311_2/conda-package-handling_1679335424584/work
conda_package_streaming @ file:///Users/ec2-user/ci_py311/conda-package-streaming_1678390826328/work
cryptography @ file:///Users/ec2-user/ci_py311_2/cryptography_1679335457662/work
filelock @ file:///Users/ec2-user/ci_py311/filelock_1678318171955/work
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
idna @ file:///Users/ec2-user/ci_py311/idna_1678315819133/work
Jinja2 @ file:///Users/ec2-user/ci_py311/jinja2_1678317138271/work
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
MarkupSafe @ file:///Users/ec2-user/ci_py311/markupsafe_1678316973889/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1724034855780/work/dist/numpy-2.1.0-cp311-cp311-macosx_10_13_x86_64.whl#sha256=90e066a54c118f4fb6cb264ee216adc7cdf10b003c50f655c4441150b4a932ac
packaging @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c3zlhpboab/croot/packaging_1678965318579/work
pkginfo @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d976p03z5u/croot/pkginfo_1679431175529/work
pluggy @ file:///Users/ec2-user/ci_py311/pluggy_1678315400971/work
protobuf==4.24.4
psutil @ file:///Users/ec2-user/ci_py311_2/psutil_1679337242143/work
pycosat @ file:///Users/ec2-user/ci_py311/pycosat_1678378899646/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_19gfn1ib_u/croot/pyopenssl_1678965300171/work
PySocks @ file:///Users/ec2-user/ci_py311/pysocks_1678315868424/work
pytz @ file:///Users/ec2-user/ci_py311/pytz_1678318034461/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1666772577178/work
requests @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ec2sxfa_zk/croot/requests_1682607519671/work
ruamel.yaml @ file:///Users/ec2-user/ci_py311/ruamel.yaml_1678379291841/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_a42j8ggi70/croot/soupsieve_1680518482618/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tomlkit @ file:///home/conda/feedstock_root/build_artifacts/tomlkit_1679924068997/work
toolz @ file:///Users/ec2-user/ci_py311/toolz_1678323009669/work
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_03acd6t6ca/croot/tqdm_1679561866522/work
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_61c6ppfd7m/croot/urllib3_1680254700568/work
xmltodict @ file:///home/conda/feedstock_root/build_artifacts/xmltodict_1652020822199/work
yq @ file:///home/conda/feedstock_root/build_artifacts/yq_1682205044037/work
zstandard @ file:///Users/ec2-user/ci_py311_2/zstandard_1679338593428/work
